Struct irc::client::conn::MockConnection

pub struct MockConnection {
    // some fields omitted
}

A mock connection for testing purposes.

Methods

impl MockConnection

fn empty() -> MockConnection

Creates a new mock connection with an empty read buffer.

fn new(input: &str) -> MockConnection

Creates a new mock connection with the specified string in the read buffer.

fn from_byte_vec(input: Vec<u8>) -> MockConnection

Creates a new mock connection with the specified bytes in the read buffer.