Enum irc::client::data::command::Command

pub enum Command {
    PASS(String),
    NICK(String),
    USER(String, String, String),
    OPER(String, String),
    MODE(String, String, Option<String>),
    SERVICE(String, String, String, String, String, String),
    QUIT(Option<String>),
    SQUIT(String, String),
    JOIN(String, Option<String>),
    PART(String, Option<String>),
    TOPIC(String, Option<String>),
    NAMES(Option<String>, Option<String>),
    LIST(Option<String>, Option<String>),
    INVITE(String, String),
    KICK(String, String, Option<String>),
    PRIVMSG(String, String),
    NOTICE(String, String),
    MOTD(Option<String>),
    LUSERS(Option<String>, Option<String>),
    VERSION(Option<String>),
    STATS(Option<String>, Option<String>),
    LINKS(Option<String>, Option<String>),
    TIME(Option<String>),
    CONNECT(String, String, Option<String>),
    TRACE(Option<String>),
    ADMIN(Option<String>),
    INFO(Option<String>),
    SERVLIST(Option<String>, Option<String>),
    SQUERY(String, String),
    WHO(Option<String>, Option<bool>),
    WHOIS(Option<String>, String),
    WHOWAS(String, Option<String>, Option<String>),
    KILL(String, String),
    PING(String, Option<String>),
    PONG(String, Option<String>),
    ERROR(String),
    AWAY(Option<String>),
    REHASH,
    DIE,
    RESTART,
    SUMMON(String, Option<String>, Option<String>),
    USERS(Option<String>),
    WALLOPS(String),
    USERHOST(Vec<String>),
    ISON(Vec<String>),
    SAJOIN(String, String),
    SAMODE(String, String, Option<String>),
    SANICK(String, String),
    SAPART(String, String),
    SAQUIT(String, String),
    NICKSERV(String),
    CHANSERV(String),
    OPERSERV(String),
    BOTSERV(String),
    HOSTSERV(String),
    MEMOSERV(String),
}

List of all client commands as defined in RFC 2812. This also includes commands from the capabilities extension. Additionally, this includes some common additional commands from popular IRCds.

Variants

PASS

PASS :password

NICK

NICK :nickname

USER

USER user mode * :realname

OPER

OPER name :password

MODE

MODE nickname modes MODE channel modes [modeparams]

SERVICE

SERVICE nickname reserved distribution type reserved :info

QUIT

QUIT :comment

SQUIT

SQUIT server :comment

JOIN

JOIN chanlist [chankeys]

PART

PART chanlist :[comment]

TOPIC

TOPIC channel :[topic]

NAMES

NAMES [chanlist :[target]]

LIST

LIST [chanlist :[target]]

INVITE

INVITE nickname channel

KICK

KICK chanlist userlist :[comment]

PRIVMSG

PRIVMSG msgtarget :message

NOTICE

NOTICE msgtarget :message

MOTD

MOTD :[target]

LUSERS

LUSERS [mask :[target]]

VERSION

VERSION :[target]

STATS

STATS [query :[target]]

LINKS

LINKS [[remote server] server :mask]

TIME

TIME :[target]

CONNECT

CONNECT target server port :[remote server]

TRACE

TRACE :[target]

ADMIN

ADMIN :[target]

INFO

INFO :[target]

SERVLIST

SERVLIST [mask :[type]]

SQUERY

SQUERY servicename text

WHO

WHO [mask ["o"]]

WHOIS

WHOIS [target] masklist

WHOWAS

WHOWAS nicklist [count :[target]]

KILL

KILL nickname :comment

PING

PING server1 :[server2]

PONG

PONG server :[server2]

ERROR

ERROR :message

AWAY

AWAY :[message]

REHASH

REHASH

DIE

DIE

RESTART

RESTART

SUMMON

SUMMON user [target :[channel]]

USERS

USERS :[target]

WALLOPS

WALLOPS :Text to be sent

USERHOST

USERHOST space-separated nicklist

ISON

ISON space-separated nicklist

SAJOIN

SAJOIN nickname channel

SAMODE

SAMODE target modes [modeparams]

SANICK

SANICK old nickname new nickname

SAPART

SAPART nickname :comment

SAQUIT

SAQUIT nickname :comment

NICKSERV

NICKSERV message

CHANSERV

CHANSERV message

OPERSERV

OPERSERV message

BOTSERV

BOTSERV message

HOSTSERV

HOSTSERV message

MEMOSERV

MEMOSERV message

Methods

impl Command

fn from_message(m: &Message) -> Result<Command>

Converts a Message into a Command.

fn from_message_io(m: Result<Message>) -> Result<Command>

Converts a potential Message result into a potential Command result.