Struct irc::client::data::message::Message

pub struct Message {
    pub tags: Option<Vec<Tag>>,
    pub prefix: Option<String>,
    pub command: String,
    pub args: Vec<String>,
    pub suffix: Option<String>,
}

IRC Message data.

Fields

tags

Message tags as defined by IRCv3.2.

prefix

The message prefix (or source) as defined by RFC 2812.

command

The IRC command as defined by RFC 2812.

args

The command arguments.

suffix

The message suffix as defined by RFC 2812. This is the only part of the message that is allowed to contain spaces.

Methods

impl Message

fn new(prefix: Option<&str>, command: &str, args: Option<Vec<&str>>, suffix: Option<&str>) -> Message

Creates a new Message.

fn with_tags(tags: Option<Vec<Tag>>, prefix: Option<&str>, command: &str, args: Option<Vec<&str>>, suffix: Option<&str>) -> Message

Creates a new Message optionally including IRCv3.2 message tags.

fn from_owned(prefix: Option<String>, command: String, args: Option<Vec<String>>, suffix: Option<String>) -> Message

Creates a new Message from already owned data.

fn get_source_nickname(&self) -> Option<&str>

Gets the nickname of the message source, if it exists.

fn into_string(&self) -> String

Converts a Message into a String according to the IRC protocol.