Struct irc::client::server::IrcServer

pub struct IrcServer<T: IrcRead, U: IrcWrite> {
    // some fields omitted
}

A thread-safe implementation of an IRC Server connection.

Methods

impl IrcServer<BufReader<NetStream>, BufWriter<NetStream>>

fn new<P: AsRef<Path>>(config: P) -> Result<NetIrcServer>

Creates a new IRC Server connection from the configuration at the specified path, connecting immediately.

fn from_config(config: Config) -> Result<NetIrcServer>

Creates a new IRC server connection from the specified configuration, connecting immediately.

fn reconnect(&self) -> Result<()>

Reconnects to the IRC server.

impl<T: IrcRead, U: IrcWrite> IrcServer<T, U>

fn from_connection(config: Config, conn: Connection<T, U>) -> IrcServer<T, U>

Creates an IRC server from the specified configuration, and any arbitrary Connection.

fn conn(&self) -> &Connection<T, U>

Gets a reference to the IRC server's connection.