Trait irc::client::server::Server

pub trait Server<'a, T, U> {
    fn config(&self) -> &Config;
    fn send(&self, command: Command) -> Result<()>;
    fn iter(&'a self) -> ServerIterator<'a, T, U>;
    fn iter_cmd(&'a self) -> ServerCmdIterator<'a, T, U>;
    fn list_users(&self, channel: &str) -> Option<Vec<User>>;
}

Trait describing core Server functionality.

Required Methods

fn config(&self) -> &Config

Gets the configuration being used with this Server.

fn send(&self, command: Command) -> Result<()>

Sends a Command to this Server.

fn iter(&'a self) -> ServerIterator<'a, T, U>

Gets an Iterator over Messages received by this Server.

fn iter_cmd(&'a self) -> ServerCmdIterator<'a, T, U>

Gets an Iterator over Commands received by this Server.

fn list_users(&self, channel: &str) -> Option<Vec<User>>

Gets a list of Users in the specified channel. This will be none if the channel is not being tracked, or if tracking is not supported altogether.