Struct irc::client::server::utils::WrapperStable

pub struct Wrapper<'a, T: IrcReader, U: IrcWriter> {
    // some fields omitted
}

Functionality-providing wrapper for Server. Wrappers are currently not thread-safe, and should be created per-thread, as needed.

Methods

impl<'a, T: IrcReader, U: IrcWriter> Wrapper<'a, T, U>

fn new(server: &'a Server<'a, T, U>) -> Wrapper<'a, T, U>

Creates a new Wrapper from the given Server.

fn identify(&self) -> IoResult<()>

Sends a NICK and USER to identify.

fn send_pong(&self, msg: &str) -> IoResult<()>

Sends a PONG with the specified message.

fn send_join(&self, chanlist: &str) -> IoResult<()>

Joins the specified channel or chanlist.

fn send_oper(&self, username: &str, password: &str) -> IoResult<()>

Attempts to oper up using the specified username and password.

fn send_privmsg(&self, target: &str, message: &str) -> IoResult<()>

Sends a message to the specified target.

fn send_notice(&self, target: &str, message: &str) -> IoResult<()>

Sends a notice to the specified target.

fn send_topic(&self, channel: &str, topic: &str) -> IoResult<()>

Sets the topic of a channel or requests the current one. If topic is an empty string, it won't be included in the message.

fn send_kill(&self, target: &str, message: &str) -> IoResult<()>

Kills the target with the provided message.

fn send_kick(&self, chanlist: &str, nicklist: &str, message: &str) -> IoResult<()>

Kicks the listed nicknames from the listed channels with a comment. If message is an empty string, it won't be included in the message.

fn send_mode(&self, target: &str, mode: &str, modeparams: &str) -> IoResult<()>

Changes the mode of the target. If modeparmas is an empty string, it won't be included in the message.

fn send_samode(&self, target: &str, mode: &str, modeparams: &str) -> IoResult<()>

Changes the mode of the target by force. If modeparams is an empty string, it won't be included in the message.

fn send_sanick(&self, old_nick: &str, new_nick: &str) -> IoResult<()>

Forces a user to change from the old nickname to the new nickname.

fn send_invite(&self, nick: &str, chan: &str) -> IoResult<()>

Invites a user to the specified channel.

fn send_quit(&self, msg: &str) -> IoResult<()>

Quits the server entirely with a message. This defaults to Powered by Rust. if none is specified.

fn send_ctcp(&self, target: &str, msg: &str) -> IoResult<()>

Sends a CTCP-escaped message to the specified target. This requires the CTCP feature to be enabled.

fn send_action(&self, target: &str, msg: &str) -> IoResult<()>

Sends an action command to the specified target. This requires the CTCP feature to be enabled.

fn send_finger(&self, target: &str) -> IoResult<()>

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_version(&self, target: &str) -> IoResult<()>

Sends a version request to the specified target. This requires the CTCP feature to be enabled.

fn send_source(&self, target: &str) -> IoResult<()>

Sends a source request to the specified target. This requires the CTCP feature to be enabled.

fn send_user_info(&self, target: &str) -> IoResult<()>

Sends a user info request to the specified target. This requires the CTCP feature to be enabled.

fn send_ctcp_ping(&self, target: &str) -> IoResult<()>

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_time(&self, target: &str) -> IoResult<()>

Sends a time request to the specified target. This requires the CTCP feature to be enabled.