Trait irc::client::server::utils::ServerExt

pub trait ServerExt<'a, T, U>: Server<'a, T, U> {
    fn identify(&self) -> Result<()> { ... }
    fn send_pong(&self, msg: &str) -> Result<()> { ... }
    fn send_join(&self, chanlist: &str) -> Result<()> { ... }
    fn send_oper(&self, username: &str, password: &str) -> Result<()> { ... }
    fn send_privmsg(&self, target: &str, message: &str) -> Result<()> { ... }
    fn send_notice(&self, target: &str, message: &str) -> Result<()> { ... }
    fn send_topic(&self, channel: &str, topic: &str) -> Result<()> { ... }
    fn send_kill(&self, target: &str, message: &str) -> Result<()> { ... }
    fn send_kick(&self, chanlist: &str, nicklist: &str, message: &str) -> Result<()> { ... }
    fn send_mode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> { ... }
    fn send_samode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> { ... }
    fn send_sanick(&self, old_nick: &str, new_nick: &str) -> Result<()> { ... }
    fn send_invite(&self, nick: &str, chan: &str) -> Result<()> { ... }
    fn send_quit(&self, msg: &str) -> Result<()> { ... }
    fn send_ctcp(&self, target: &str, msg: &str) -> Result<()> { ... }
    fn send_action(&self, target: &str, msg: &str) -> Result<()> { ... }
    fn send_finger(&self, target: &str) -> Result<()> { ... }
    fn send_version(&self, target: &str) -> Result<()> { ... }
    fn send_source(&self, target: &str) -> Result<()> { ... }
    fn send_user_info(&self, target: &str) -> Result<()> { ... }
    fn send_ctcp_ping(&self, target: &str) -> Result<()> { ... }
    fn send_time(&self, target: &str) -> Result<()> { ... }
}

Extensions for Server capabilities that make it easier to work directly with the protocol.

Provided Methods

fn identify(&self) -> Result<()>

Sends a NICK and USER to identify.

fn send_pong(&self, msg: &str) -> Result<()>

Sends a PONG with the specified message.

fn send_join(&self, chanlist: &str) -> Result<()>

Joins the specified channel or chanlist.

fn send_oper(&self, username: &str, password: &str) -> Result<()>

Attempts to oper up using the specified username and password.

fn send_privmsg(&self, target: &str, message: &str) -> Result<()>

Sends a message to the specified target.

fn send_notice(&self, target: &str, message: &str) -> Result<()>

Sends a notice to the specified target.

fn send_topic(&self, channel: &str, topic: &str) -> Result<()>

Sets the topic of a channel or requests the current one. If topic is an empty string, it won't be included in the message.

fn send_kill(&self, target: &str, message: &str) -> Result<()>

Kills the target with the provided message.

fn send_kick(&self, chanlist: &str, nicklist: &str, message: &str) -> Result<()>

Kicks the listed nicknames from the listed channels with a comment. If message is an empty string, it won't be included in the message.

fn send_mode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()>

Changes the mode of the target. If modeparmas is an empty string, it won't be included in the message.

fn send_samode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()>

Changes the mode of the target by force. If modeparams is an empty string, it won't be included in the message.

fn send_sanick(&self, old_nick: &str, new_nick: &str) -> Result<()>

Forces a user to change from the old nickname to the new nickname.

fn send_invite(&self, nick: &str, chan: &str) -> Result<()>

Invites a user to the specified channel.

fn send_quit(&self, msg: &str) -> Result<()>

Quits the server entirely with a message. This defaults to Powered by Rust. if none is specified.

fn send_ctcp(&self, target: &str, msg: &str) -> Result<()>

Sends a CTCP-escaped message to the specified target. This requires the CTCP feature to be enabled.

fn send_action(&self, target: &str, msg: &str) -> Result<()>

Sends an action command to the specified target. This requires the CTCP feature to be enabled.

fn send_finger(&self, target: &str) -> Result<()>

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_version(&self, target: &str) -> Result<()>

Sends a version request to the specified target. This requires the CTCP feature to be enabled.

fn send_source(&self, target: &str) -> Result<()>

Sends a source request to the specified target. This requires the CTCP feature to be enabled.

fn send_user_info(&self, target: &str) -> Result<()>

Sends a user info request to the specified target. This requires the CTCP feature to be enabled.

fn send_ctcp_ping(&self, target: &str) -> Result<()>

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_time(&self, target: &str) -> Result<()>

Sends a time request to the specified target. This requires the CTCP feature to be enabled.