Enum irc::data::command::CapSubCommandStable

pub enum CapSubCommand {
    LS,
    LIST,
    REQ,
    ACK,
    NAK,
    CLEAR,
    END,
}

A list of all of the subcommands for the capabilities extension.

Variants

LS

Requests a list of the server's capabilities.

LIST

Requests a list of the server's capabilities.

REQ

Requests specific capabilities blindly.

ACK

Acknowledges capabilities.

NAK

Does not acknowledge certain capabilities.

CLEAR

Requests that the server clears the capabilities of this client.

END

Ends the capability negotiation before registration.

Methods

impl CapSubCommand

fn to_str(&self) -> &str

Gets the string that corresponds to this subcommand.