Enum irc::data::command::CommandStable

pub enum Command<'a> {
    PASS(&'a str),
    NICK(&'a str),
    USER(&'a str, &'a str, &'a str),
    OPER(&'a str, &'a str),
    MODE(&'a str, &'a str, Option<&'a str>),
    SERVICE(&'a str, &'a str, &'a str, &'a str, &'a str, &'a str),
    QUIT(Option<&'a str>),
    SQUIT(&'a str, &'a str),
    JOIN(&'a str, Option<&'a str>),
    PART(&'a str, Option<&'a str>),
    TOPIC(&'a str, Option<&'a str>),
    NAMES(Option<&'a str>, Option<&'a str>),
    LIST(Option<&'a str>, Option<&'a str>),
    INVITE(&'a str, &'a str),
    KICK(&'a str, &'a str, Option<&'a str>),
    PRIVMSG(&'a str, &'a str),
    NOTICE(&'a str, &'a str),
    MOTD(Option<&'a str>),
    LUSERS(Option<&'a str>, Option<&'a str>),
    VERSION(Option<&'a str>),
    STATS(Option<&'a str>, Option<&'a str>),
    LINKS(Option<&'a str>, Option<&'a str>),
    TIME(Option<&'a str>),
    CONNECT(&'a str, &'a str, Option<&'a str>),
    TRACE(Option<&'a str>),
    ADMIN(Option<&'a str>),
    INFO(Option<&'a str>),
    SERVLIST(Option<&'a str>, Option<&'a str>),
    SQUERY(&'a str, &'a str),
    WHO(Option<&'a str>, Option<bool>),
    WHOIS(Option<&'a str>, &'a str),
    WHOWAS(&'a str, Option<&'a str>, Option<&'a str>),
    KILL(&'a str, &'a str),
    PING(&'a str, Option<&'a str>),
    PONG(&'a str, Option<&'a str>),
    ERROR(&'a str),
    AWAY(Option<&'a str>),
    REHASH,
    DIE,
    RESTART,
    SUMMON(&'a str, Option<&'a str>, Option<&'a str>),
    USERS(Option<&'a str>),
    WALLOPS(&'a str),
    USERHOST(Vec<&'a str>),
    ISON(Vec<&'a str>),
    SAJOIN(&'a str, &'a str),
    SAMODE(&'a str, &'a str, Option<&'a str>),
    SANICK(&'a str, &'a str),
    SAPART(&'a str, &'a str),
    SAQUIT(&'a str, &'a str),
    NICKSERV(&'a str),
    CHANSERV(&'a str),
    OPERSERV(&'a str),
    BOTSERV(&'a str),
    HOSTSERV(&'a str),
    MEMOSERV(&'a str),
    CAP(CapSubCommand, Option<&'a str>),
}

List of all client commands as defined in RFC 2812. This also includes commands from the capabilities extension. Additionally, this includes some common additional commands from popular IRCds.

Variants

PASS

PASS password

NICK

NICK nickname

USER

USER user mode * realname

OPER

OPER name password

MODE

MODE nickname modes MODE channel modes [modeparams]

SERVICE

SERVICE nickname reserved distribution type reserved info

QUIT

QUIT Quit Message

SQUIT

SQUIT server comment

JOIN

JOIN chanlist [chankeys]

PART

PART chanlist [Part Message]

TOPIC

TOPIC channel [topic]

NAMES

NAMES [chanlist [target]]

LIST

LIST [chanlist [target]]

INVITE

INVITE nickname channel

KICK

KICK chanlist userlist [comment]

PRIVMSG

PRIVMSG msgtarget text to be sent

NOTICE

NOTICE msgtarget text

MOTD

MOTD [target]

LUSERS

LUSERS [mask [target]]

VERSION

VERSION [target]

STATS

STATS [query [target]]

LINKS

LINKS [[remote server] server mask]

TIME

TIME [target]

CONNECT

CONNECT target server port [remote server]

TRACE

TRACE [target]

ADMIN

ADMIN [target]

INFO

INFO [target]

SERVLIST

SERVLIST [mask [type]]

SQUERY

SQUERY servicename text

WHO

WHO [mask ["o"]]

WHOIS

WHOIS [target] masklist

WHOWAS

WHOWAS nicklist [count [target]]

KILL

KILL nickname comment

PING

PING server1 [server2]

PONG

PONG server [server2]

ERROR

ERROR error message

AWAY

AWAY [text]

REHASH

REHASH

DIE

DIE

RESTART

RESTART

SUMMON

SUMMON user [target [channel]]

USERS

USERS [target]

WALLOPS

WALLOPS Text to be sent

USERHOST

USERHOST space-separated nicklist

ISON

ISON space-separated nicklist

SAJOIN

SAJOIN nickname channel

SAMODE

SAMODE target modes [modeparams]

SANICK

SANICK old nickname new nickname

SAPART

SAPART nickname reason

SAQUIT

SAQUIT nickname reason

NICKSERV

NICKSERV message

CHANSERV

CHANSERV message

OPERSERV

OPERSERV message

BOTSERV

BOTSERV message

HOSTSERV

HOSTSERV message

MEMOSERV

MEMOSERV message

CAP

CAP COMMAND [param]

Methods

impl<'a> Command<'a>

fn to_message(self) -> Message

Converts a Command into a Message.

fn from_message(m: &'a Message) -> IoResult<Command<'a>>

Converts a Message into a Command.