Enum irc::data::response::ResponseUnstable

pub enum Response {
    RPL_WELCOME,
    RPL_YOURHOST,
    RPL_CREATED,
    RPL_MYINFO,
    RPL_BOUNCE,
    RPL_USERHOST,
    RPL_ISON,
    RPL_AWAY,
    RPL_UNAWAY,
    RPL_NOWAWAY,
    RPL_WHOISUSER,
    RPL_WHOISSERVER,
    RPL_WHOISOPERATOR,
    RPL_WHOISIDLE,
    RPL_ENDOFWHOIS,
    RPL_WHOISCHANNELS,
    RPL_WHOWASUSER,
    RPL_ENDOFWHOWAS,
    RPL_LISTSTART,
    RPL_LIST,
    RPL_LISTEND,
    RPL_UNIQOPIS,
    RPL_CHANNELMODEIS,
    RPL_NOTOPIC,
    RPL_TOPIC,
    RPL_INVITING,
    RPL_SUMMONING,
    RPL_INVITELIST,
    RPL_ENDOFINVITELIST,
    RPL_EXCEPTLIST,
    RPL_ENDOFEXECPTLIST,
    RPL_VERSION,
    RPL_WHOREPLY,
    RPL_ENDOFWHO,
    RPL_NAMREPLY,
    RPL_ENDOFNAMES,
    RPL_LINKS,
    RPL_ENDOFLINKS,
    RPL_BANLIST,
    RPL_ENDOFBANLIST,
    RPL_INFO,
    RPL_ENDOFINFO,
    RPL_MOTDSTART,
    RPL_MOTD,
    RPL_ENDOFMOTD,
    RPL_YOUREOPER,
    RPL_REHASHING,
    RPL_YOURESERVICE,
    RPL_TIME,
    RPL_USERSSTART,
    RPL_USERS,
    RPL_ENDOFUSERS,
    RPL_NOUSERS,
    RPL_TRACELINK,
    RPL_TRACECONNECTING,
    RPL_TRACEHANDSHAKE,
    RPL_TRACEUKNOWN,
    RPL_TRACEOPERATOR,
    RPL_TRACEUSER,
    RPL_TRACESERVER,
    RPL_TRACESERVICE,
    RPL_TRACENEWTYPE,
    RPL_TRACECLASS,
    RPL_TRACERECONNECT,
    RPL_TRACELOG,
    RPL_TRACEEND,
    RPL_STATSLINKINFO,
    RPL_STATSCOMMANDS,
    RPL_ENDOFSTATS,
    RPL_STATSUPTIME,
    RPL_STATSOLINE,
    RPL_UMODEIS,
    RPL_SERVLIST,
    RPL_SERVLISTEND,
    RPL_LUSERCLIENT,
    RPL_LUSEROP,
    RPL_LUSERUNKNOWN,
    RPL_LUSERCHANNELS,
    RPL_LUSERME,
    RPL_ADMINME,
    RPL_ADMINLOC1,
    RPL_ADMINLOC2,
    RPL_ADMINEMAIL,
    RPL_TRYAGAIN,
    ERR_NOSUCHNICK,
    ERR_NOSUCHSERVER,
    ERR_NOSUCHCHANNEL,
    ERR_CANNOTSENDTOCHAN,
    ERR_TOOMANYCHANNELS,
    ERR_WASNOSUCHNICK,
    ERR_TOOMANYTARGETS,
    ERR_NOSUCHSERVICE,
    ERR_NOORIGIN,
    ERR_NORECIPIENT,
    ERR_NOTEXTTOSEND,
    ERR_NOTOPLEVEL,
    ERR_WILDTOPLEVEL,
    ERR_BADMASK,
    ERR_UNKNOWNCOMMAND,
    ERR_NOMOTD,
    ERR_NOADMININFO,
    ERR_FILEERROR,
    ERR_NONICKNAMEGIVEN,
    ERR_ERRONEOUSNICKNAME,
    ERR_NICKNAMEINUSE,
    ERR_NICKCOLLISION,
    ERR_UNAVAILRESOURCE,
    ERR_USERNOTINCHANNEL,
    ERR_NOTONCHANNEL,
    ERR_USERONCHANNEL,
    ERR_NOLOGIN,
    ERR_SUMMONDISABLED,
    ERR_USERSDISABLED,
    ERR_NOTREGISTERED,
    ERR_NEEDMOREPARAMS,
    ERR_ALREADYREGISTRED,
    ERR_NOPERMFORHOST,
    ERR_PASSWDMISMATCH,
    ERR_YOUREBANNEDCREEP,
    ERR_YOUWILLBEBANNED,
    ERR_KEYSET,
    ERR_CHANNELISFULL,
    ERR_UNKNOWNMODE,
    ERR_INVITEONLYCHAN,
    ERR_BANNEDFROMCHAN,
    ERR_BADCHANNELKEY,
    ERR_BADCHANMASK,
    ERR_NOCHANMODES,
    ERR_BANLISTFULL,
    ERR_NOPRIVILEGES,
    ERR_CHANOPRIVSNEEDED,
    ERR_CANTKILLSERVER,
    ERR_RESTRICTED,
    ERR_UNIQOPPRIVSNEEDED,
    ERR_NOOPERHOST,
    ERR_UMODEUNKNOWNFLAG,
    ERR_USERSDONTMATCH,
}

List of all server responses as defined in RFC 2812. All commands are documented with their expected form from the RFC.

Variants

RPL_WELCOME

001 Welcome to the Internet Relay Network !@

RPL_YOURHOST

002 Your host is , running version

RPL_CREATED

003 This server was created

RPL_MYINFO

004 available channel modes>

RPL_BOUNCE

005 Try server , port

RPL_USERHOST

302 :*1 *( " " )

RPL_ISON

303 :*1 *( " " )

RPL_AWAY

301 :

RPL_UNAWAY

305 :You are no longer marked as being away

RPL_NOWAWAY

306 :You have been marked as being away

RPL_WHOISUSER

311 * :

RPL_WHOISSERVER

312 :

RPL_WHOISOPERATOR

313 :is an IRC operator

RPL_WHOISIDLE

317 :seconds idle

RPL_ENDOFWHOIS

318 :End of WHOIS list

RPL_WHOISCHANNELS

319 :*( ( "@" / "+" ) " " )

RPL_WHOWASUSER

314 * :

RPL_ENDOFWHOWAS

369 :End of WHOWAS

RPL_LISTSTART

Obsolete. Not used.

RPL_LIST

322 <# visible> :

RPL_LISTEND

323 :End of LIST

RPL_UNIQOPIS

325

RPL_CHANNELMODEIS

324

RPL_NOTOPIC

331 :No topic is set

RPL_TOPIC

332 :

RPL_INVITING

341

RPL_SUMMONING

342 :Summoning user to IRC

RPL_INVITELIST

346

RPL_ENDOFINVITELIST

347 :End of channel invite list

RPL_EXCEPTLIST

348

RPL_ENDOFEXECPTLIST

349 :End of channel exception list

RPL_VERSION

351 . :

RPL_WHOREPLY

352 ( "H" / "G" > ["*"] [ ( "@" / "+" ) ] :

RPL_ENDOFWHO

315 :End of WHO list

RPL_NAMREPLY

353 ( "=" / "*" / "@" ) :[ "@" / "+" ] *( " " [ "@" / "+" ] )

RPL_ENDOFNAMES

366 :End of NAMES list

RPL_LINKS

364 :

RPL_ENDOFLINKS

365 :End of LINKS list

RPL_BANLIST

367

RPL_ENDOFBANLIST

368 :End of channel ban list

RPL_INFO

371 :

RPL_ENDOFINFO

374 :End of INFO list

RPL_MOTDSTART

375 :- Message of the day -

RPL_MOTD

372 :-

RPL_ENDOFMOTD

376 :End of MOTD command

RPL_YOUREOPER

381 :You are now an IRC operator

RPL_REHASHING

382 :Rehashing

RPL_YOURESERVICE

383 You are service

RPL_TIME

391 :

RPL_USERSSTART

392 :UserID Terminal Host

RPL_USERS

393 :

RPL_ENDOFUSERS

394 :End of users

RPL_NOUSERS

395 :Nobody logged in

RPL_TRACELINK

200 Link V

RPL_TRACECONNECTING

201 Try.

RPL_TRACEHANDSHAKE

202 H.S.

RPL_TRACEUKNOWN

203 ???? []

RPL_TRACEOPERATOR

204 Oper

RPL_TRACEUSER

205 User

RPL_TRACESERVER

206 Serv S C @ V

RPL_TRACESERVICE

207 Service

RPL_TRACENEWTYPE

208 0

RPL_TRACECLASS

209 Class

RPL_TRACERECONNECT

Unused.

RPL_TRACELOG

261 File

RPL_TRACEEND

262 :End of TRACE

RPL_STATSLINKINFO

211

RPL_STATSCOMMANDS

212

RPL_ENDOFSTATS

219 :End of STATS report

RPL_STATSUPTIME

242 :Server Up %d days %d:%02d:%02d

RPL_STATSOLINE

O *

RPL_UMODEIS

221

RPL_SERVLIST

234

RPL_SERVLISTEND

235 :End of service listing

RPL_LUSERCLIENT

251 :There are users and services on servers

RPL_LUSEROP

252 :operator(s) online

RPL_LUSERUNKNOWN

253 :unknown connection(s)

RPL_LUSERCHANNELS

254 :channels formed

RPL_LUSERME

255 :I have clients and servers

RPL_ADMINME

256 :Administrative info

RPL_ADMINLOC1

257 :

RPL_ADMINLOC2

258 :

RPL_ADMINEMAIL

259 :

RPL_TRYAGAIN

263 :Please wait a while and try again.

ERR_NOSUCHNICK

401 :No such nick/channel

ERR_NOSUCHSERVER

402 :No such server

ERR_NOSUCHCHANNEL

403 :No such channel

ERR_CANNOTSENDTOCHAN

404 :Cannot send to channel

ERR_TOOMANYCHANNELS

405 :You have joined too many channels

ERR_WASNOSUCHNICK

406 :There was no such nickname

ERR_TOOMANYTARGETS

407 : recipients.

ERR_NOSUCHSERVICE

408 :No such service

ERR_NOORIGIN

409 :No origin specified

ERR_NORECIPIENT

411 :No recipient given ()

ERR_NOTEXTTOSEND

412 :No text to send

ERR_NOTOPLEVEL

413 :No toplevel domain specified

ERR_WILDTOPLEVEL

414 :Wildcard in toplevel domain

ERR_BADMASK

415 :Bad Server/host mask

ERR_UNKNOWNCOMMAND

421 :Unknown command

ERR_NOMOTD

422 :MOTD File is missing

ERR_NOADMININFO

423 :No administrative info available

ERR_FILEERROR

424 :File error doing on

ERR_NONICKNAMEGIVEN

431 :No nickname given

ERR_ERRONEOUSNICKNAME

432 :Erroneous nickname"

ERR_NICKNAMEINUSE

433 :Nickname is already in use

ERR_NICKCOLLISION

436 :Nickname collision KILL from @

ERR_UNAVAILRESOURCE

437 :Nick/channel is temporarily unavailable

ERR_USERNOTINCHANNEL

441 :They aren't on that channel

ERR_NOTONCHANNEL

442 :You're not on that channel

ERR_USERONCHANNEL

443 :is already on channel

ERR_NOLOGIN

444 :User not logged in

ERR_SUMMONDISABLED

445 :SUMMON has been disabled

ERR_USERSDISABLED

446 :USERS has been disabled

ERR_NOTREGISTERED

451 :You have not registered

ERR_NEEDMOREPARAMS

461 :Not enough parameters

ERR_ALREADYREGISTRED

462 :Unauthorized command (already registered)

ERR_NOPERMFORHOST

463 :Your host isn't among the privileged

ERR_PASSWDMISMATCH

464 :Password incorrect

ERR_YOUREBANNEDCREEP

465 :You are banned from this server

ERR_YOUWILLBEBANNED

466

ERR_KEYSET

467 :Channel key already set

ERR_CHANNELISFULL

471 :Cannot join channel (+l)

ERR_UNKNOWNMODE

472 :is unknown mode char to me for

ERR_INVITEONLYCHAN

473 :Cannot join channel (+i)

ERR_BANNEDFROMCHAN

474 :Cannot join channel (+b)

ERR_BADCHANNELKEY

475 :Cannot join channel (+k)

ERR_BADCHANMASK

476 :Bad Channel Mask

ERR_NOCHANMODES

477 :Channel doesn't support modes

ERR_BANLISTFULL

478 :Channel list is full

ERR_NOPRIVILEGES

481 :Permission Denied- You're not an IRC operator

ERR_CHANOPRIVSNEEDED

482 :You're not channel operator

ERR_CANTKILLSERVER

483 :You can't kill a server!

ERR_RESTRICTED

484 :Your connection is restricted!

ERR_UNIQOPPRIVSNEEDED

485 :You're not the original channel operator

ERR_NOOPERHOST

491 :No O-lines for your host

ERR_UMODEUNKNOWNFLAG

501 :Unknown MODE flag

ERR_USERSDONTMATCH

502 :Cannot change mode for other users

Methods

impl Response

fn from_message(m: &Message) -> Option<Response>

Gets a response from a message.

fn is_error(&self) -> bool

Determines whether or not this response is an error response.