Module irc::event_stream

Structs

EventStream
EventStream wraps a stream and will loop, reading lines and processing them with Handlers, which are allowed to write back to the stream and install additional Handlers.

Response
Optionally respond with a message, then take the given next Action.

Enums

Action
What should we do next?

HandlerAction

Type Definitions

Handler
A Handler examines a line of input read from a stream, and produces an optional Response to send, an additional Handler to install (perhaps to wait for a response), and a next Action to take after processing the line.