Struct irc::event_stream::EventStream

pub struct EventStream {
    // some fields omitted
}

EventStream wraps a stream and will loop, reading lines and processing them with Handlers, which are allowed to write back to the stream and install additional Handlers.

It implements Write so you can still write manually, and you can install additional Handlers when you want.

Methods

impl EventStream

fn new<R: Read + Send + 'static, W: Write + Send + 'static>(inner_reader: R, inner_writer: W, init_handlers: Vec<Handler>) -> Result<(EventStream, JoinHandle)>

Creates a new event stream with initial handlers. You should pass the Read and Write parts separately, for example using TcpStream::try_clone.

fn add_handler(&mut self, handler: Handler)

fn await_lines(&mut self, expectations: Vec<Regex>) -> Vec<Future<String>>