irc::event_stream::Handler

type Handler = Box<FnMut(&str) -> Response + Send>;

A Handler examines a line of input read from a stream, and produces an optional Response to send, an additional Handler to install (perhaps to wait for a response), and a next Action to take after processing the line.