Module irc::parseExperimental

Experimental parsing code

Structs

IrcMsg
Represents any syntactically valid IRC message. No semantic checking is applied.

IrcMsgNew

IrcMsgPrefix
An IRC prefix, which identifies the source of a message.

Functions

can_target_channel
Whether or not a command name is allowed to target a channel

is_channel
Determines whether or not an identifier is a channel, by checking the first character.

is_full_prefix
Checks whether a prefix contains nick, username and host or not.