Struct irc::parse::IrcMsgExperimental

pub struct IrcMsg<'a> {
    // some fields omitted
}

Represents any syntactically valid IRC message. No semantic checking is applied.

Methods

impl<'a> IrcMsg<'a>

fn new(msg_text: MaybeOwned<'a>) -> Option<IrcMsg<'a>>

Parses a string into an optional IrcMsg. If the string is not syntactically valid, None is returned.

fn get_prefix_raw(&'a self) -> Option<&'a str>

The prefix which identifies the source of a message, as a string slice

fn get_prefix(&'a self) -> Option<IrcMsgPrefix<'a>>

The prefix which identifies the source of a message

fn get_command(&'a self) -> &'a str

The command name of a message as received i.e. without normalisation.

fn get_args(&'a self) -> Vec<&'a str>

The arguments of a message

fn source_nick(&'a self) -> Option<&'a str>

The nick of the user a message came from, if any.