Trait irc::server::ServerExperimental

pub trait Server<'a, T, U> {
    fn config(&self) -> &Config;
    fn send(&self, _: Command) -> IoResult<()>;
    fn iter(&'a self) -> ServerIterator<'a, T, U>;
    fn list_users(&self, _: &str) -> Option<Vec<User>>;
}

Trait describing core Server functionality.

Required Methods

fn config(&self) -> &Config

Gets the configuration being used with this Server.

fn send(&self, _: Command) -> IoResult<()>

Sends a Command to this Server.

fn iter(&'a self) -> ServerIterator<'a, T, U>

Gets an Iterator over Messages received by this Server.

fn list_users(&self, _: &str) -> Option<Vec<User>>

Gets a list of Users in the specified channel.