Struct irc::Client

pub struct Client {
    // some fields omitted
}

The top-level IRC client.

Methods

impl Client

fn connect<A: ToSocketAddrs + Debug>(addr: &A, nick: &str, channels: &[&str], user: &str, realname: &str) -> Result<(Client, JoinHandle)>

Connects to an IRC server.

Returns a Client object to which you can add handlers, and a thread::JoinHandle which will join when the thread handling IRC events finishes.

Example:

let host = "irc.freenode.net";
let port = 6667;
let nick = "rustbot_test";
let channels = ["#rustbot_test"];
let (mut client, join_handle) = irc::Client::connect(
    &(host, port), nick, &channels, "username", "realname").ok().unwrap();

fn connect_mode<A: ToSocketAddrs + Debug>(addr: &A, nick: &str, channels: &[&str], user: &str, realname: &str, invisible: bool, wallops: bool) -> Result<(Client, JoinHandle)>

Connects to an IRC server with additional options invisible and wallops.

See Client::connect and the IRC protocol RFC for details.

fn add_handler(&mut self, handler: Handler)

Adds a new handler to the event loop.

Example:

use irc::event_stream::Response;

let (mut client, _) = irc::Client::connect(
    &("irc.freenode.net", 6667), "rustbot", &[], "rustbot", "rustbot").ok().unwrap();
client.add_handler(Box::new(move |line: &str| {
    // A literal echo server like this would really confuse a real IRC server...
    Response::respond(line.to_string())
}));