Struct irc::client::data::config::Config

pub struct Config {
    pub owners: Option<Vec<String>>,
    pub nickname: Option<String>,
    pub nick_password: Option<String>,
    pub alt_nicks: Option<Vec<String>>,
    pub username: Option<String>,
    pub realname: Option<String>,
    pub server: Option<String>,
    pub port: Option<u16>,
    pub password: Option<String>,
    pub use_ssl: Option<bool>,
    pub encoding: Option<String>,
    pub channels: Option<Vec<String>>,
    pub umodes: Option<String>,
    pub user_info: Option<String>,
    pub options: Option<HashMap<String, String>>,
}

Configuration data.

Fields

owners

A list of the owners of the bot by nickname.

nickname

The bot's nickname.

nick_password

The bot's NICKSERV password.

alt_nicks

Alternative nicknames for the bots, if the default is taken.

username

The bot's username.

realname

The bot's real name.

server

The server to connect to.

port

The port to connect on.

password

The password to connect to the server.

use_ssl

Whether or not to use SSL. Bots will automatically panic if this is enabled without SSL support.

encoding

The encoding type used for this connection. This is typically UTF-8, but could be something else.

channels

A list of channels to join on connection.

umodes

User modes to set on connect. Example: "+RB-x"

user_info

The text that'll be sent in response to CTCP USERINFO requests.

options

A map of additional options to be stored in config.

Methods

impl Config

fn load(path: Path) -> IoResult<Config>

Loads a JSON configuration from the desired path.

fn load_utf8(path: &str) -> IoResult<Config>

Loads a JSON configuration using the string as a UTF-8 path.

fn is_owner(&self, nickname: &str) -> bool

Determines whether or not the nickname provided is the owner of the bot.

fn nickname(&self) -> &str

Gets the nickname specified in the configuration. This will panic if not specified.

fn nick_password(&self) -> &str

Gets the bot's nickserv password specified in the configuration. This defaults to an empty string when not specified.

fn get_alternate_nicknames(&self) -> Vec<&str>

Gets the alternate nicknames specified in the configuration. This defaults to an empty vector when not specified.

fn username(&self) -> &str

Gets the username specified in the configuration. This defaults to the user's nickname when not specified.

fn real_name(&self) -> &str

Gets the real name specified in the configuration. This defaults to the user's nickname when not specified.

fn server(&self) -> &str

Gets the address of the server specified in the configuration. This panics when not specified.

fn port(&self) -> u16

Gets the port of the server specified in the configuration. This defaults to 6667 (or 6697 if use_ssl is specified as true) when not specified.

fn password(&self) -> &str

Gets the server password specified in the configuration. This defaults to a blank string when not specified.

fn use_ssl(&self) -> bool

Gets whether or not to use SSL with this connection. This defaults to false when not specified.

fn encoding(&self) -> &str

Gets the encoding to use for this connection. This requires the encode feature to work. This defaults to UTF-8 when not specified.

fn channels(&self) -> Vec<&str>

Gets the channels to join upon connection. This defaults to an empty vector if it's not specified.

fn umodes(&self) -> &str

Gets the user modes to set on connect specified in the configuration. This defaults to an empty string when not specified.

fn user_info(&self) -> &str

Gets the string to be sent in response to CTCP USERINFO requests. This defaults to an empty string when not specified.

fn get_option(&self, option: &str) -> &str

Looks up the specified string in the options map. This uses indexing, and thus panics when the string is not present. This will also panic if used and there are no options.