Struct irc::client::data::message::Message

pub struct Message {
    pub tags: Option<Vec<Tag>>,
    pub prefix: Option<String>,
    pub command: Command,
}

IRC Message data.

Fields

tags

Message tags as defined by IRCv3.2.

prefix

The message prefix (or source) as defined by RFC 2812.

command

The IRC command.

Methods

impl Message

fn new(prefix: Option<&str>, command: &str, args: Vec<&str>, suffix: Option<&str>) -> IoResult<Message>

Creates a new Message.

fn with_tags(tags: Option<Vec<Tag>>, prefix: Option<&str>, command: &str, args: Vec<&str>, suffix: Option<&str>) -> IoResult<Message>

Creates a new Message optionally including IRCv3.2 message tags.

fn source_nickname(&self) -> Option<&str>

Gets the nickname of the message source, if it exists.

fn to_string(&self) -> String

Converts a Message into a String according to the IRC protocol.