Struct irc::client::server::IrcServer

pub struct IrcServer {
    // some fields omitted
}

A thread-safe implementation of an IRC Server connection.

Methods

impl IrcServer

fn new<P: AsRef<Path>>(config: P) -> Result<IrcServer>

Creates a new IRC Server connection from the configuration at the specified path, connecting immediately.

fn from_config(config: Config) -> Result<IrcServer>

Creates a new IRC server connection from the specified configuration, connecting immediately.

impl IrcServer

fn from_connection<C>(config: Config, conn: C) -> IrcServer where C: Connection + Send + Sync + 'static

Creates an IRC server from the specified configuration, and any arbitrary sync connection.

fn conn(&self) -> &Box<Connection + Send + Sync>

Gets a reference to the IRC server's connection.

fn reconnect(&self) -> Result<()>

Reconnects to the IRC server, disconnecting if necessary.

fn current_nickname(&self) -> &str

Gets the current nickname in use.