Trait irc::client::server::Server

pub trait Server {
    fn config(&self) -> &Config;
    fn send<M: Into<Message>>(&self, message: M) -> Result<()> where Self: Sized;
    fn iter<'a>(&'a self) -> Box<Iterator<Item=Result<Message>> + 'a>;
    fn list_users(&self, channel: &str) -> Option<Vec<User>>;
}

An interface for interacting with an IRC server.

Required Methods

fn config(&self) -> &Config

Gets the configuration being used with this Server.

fn send<M: Into<Message>>(&self, message: M) -> Result<()> where Self: Sized

Sends a Command to this Server.

fn iter<'a>(&'a self) -> Box<Iterator<Item=Result<Message>> + 'a>

Gets an iterator over received messages.

fn list_users(&self, channel: &str) -> Option<Vec<User>>

Gets a list of Users in the specified channel. This will be none if the channel is not being tracked, or if tracking is not supported altogether. For best results, be sure to request multi-prefix support from the server.