Trait irc::client::server::utils::ServerExt

pub trait ServerExt: Server {
    fn send_cap_ls(&self, version: NegotiationVersion) -> Result<()> where Self: Sized { ... }
    fn send_cap_req(&self, extensions: &[Capability]) -> Result<()> where Self: Sized { ... }
    fn identify(&self) -> Result<()> where Self: Sized { ... }
    fn send_sasl(&self, data: &str) -> Result<()> where Self: Sized { ... }
    fn send_sasl_plain(&self) -> Result<()> where Self: Sized { ... }
    fn send_sasl_external(&self) -> Result<()> where Self: Sized { ... }
    fn send_sasl_abort(&self) -> Result<()> where Self: Sized { ... }
    fn send_pong(&self, msg: &str) -> Result<()> where Self: Sized { ... }
    fn send_join(&self, chanlist: &str) -> Result<()> where Self: Sized { ... }
    fn send_join_with_keys(&self, chanlist: &str, keylist: &str) -> Result<()> where Self: Sized { ... }
    fn send_oper(&self, username: &str, password: &str) -> Result<()> where Self: Sized { ... }
    fn send_privmsg(&self, target: &str, message: &str) -> Result<()> where Self: Sized { ... }
    fn send_notice(&self, target: &str, message: &str) -> Result<()> where Self: Sized { ... }
    fn send_topic(&self, channel: &str, topic: &str) -> Result<()> where Self: Sized { ... }
    fn send_kill(&self, target: &str, message: &str) -> Result<()> where Self: Sized { ... }
    fn send_kick(&self, chanlist: &str, nicklist: &str, message: &str) -> Result<()> where Self: Sized { ... }
    fn send_mode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> where Self: Sized { ... }
    fn send_samode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> where Self: Sized { ... }
    fn send_sanick(&self, old_nick: &str, new_nick: &str) -> Result<()> where Self: Sized { ... }
    fn send_invite(&self, nick: &str, chan: &str) -> Result<()> where Self: Sized { ... }
    fn send_quit(&self, msg: &str) -> Result<()> where Self: Sized { ... }
    fn send_ctcp(&self, target: &str, msg: &str) -> Result<()> where Self: Sized { ... }
    fn send_action(&self, target: &str, msg: &str) -> Result<()> where Self: Sized { ... }
    fn send_finger(&self, target: &str) -> Result<()> where Self: Sized { ... }
    fn send_version(&self, target: &str) -> Result<()> where Self: Sized { ... }
    fn send_source(&self, target: &str) -> Result<()> where Self: Sized { ... }
    fn send_user_info(&self, target: &str) -> Result<()> where Self: Sized { ... }
    fn send_ctcp_ping(&self, target: &str) -> Result<()> where Self: Sized { ... }
    fn send_time(&self, target: &str) -> Result<()> where Self: Sized { ... }
}

Extensions for Server capabilities that make it easier to work directly with the protocol.

Provided Methods

fn send_cap_ls(&self, version: NegotiationVersion) -> Result<()> where Self: Sized

Sends a request for a list of server capabilities for a specific IRCv3 version.

fn send_cap_req(&self, extensions: &[Capability]) -> Result<()> where Self: Sized

Sends an IRCv3 capabilities request for the specified extensions.

fn identify(&self) -> Result<()> where Self: Sized

Sends a CAP END, NICK and USER to identify.

fn send_sasl(&self, data: &str) -> Result<()> where Self: Sized

Sends a SASL AUTHENTICATE message with the specified data.

fn send_sasl_plain(&self) -> Result<()> where Self: Sized

Sends a SASL AUTHENTICATE request to use the PLAIN mechanism.

fn send_sasl_external(&self) -> Result<()> where Self: Sized

Sends a SASL AUTHENTICATE request to use the EXTERNAL mechanism.

fn send_sasl_abort(&self) -> Result<()> where Self: Sized

Sends a SASL AUTHENTICATE request to abort authentication.

fn send_pong(&self, msg: &str) -> Result<()> where Self: Sized

Sends a PONG with the specified message.

fn send_join(&self, chanlist: &str) -> Result<()> where Self: Sized

Joins the specified channel or chanlist.

fn send_join_with_keys(&self, chanlist: &str, keylist: &str) -> Result<()> where Self: Sized

Joins the specified channel or chanlist using the specified key or keylist.

fn send_oper(&self, username: &str, password: &str) -> Result<()> where Self: Sized

Attempts to oper up using the specified username and password.

fn send_privmsg(&self, target: &str, message: &str) -> Result<()> where Self: Sized

Sends a message to the specified target.

fn send_notice(&self, target: &str, message: &str) -> Result<()> where Self: Sized

Sends a notice to the specified target.

fn send_topic(&self, channel: &str, topic: &str) -> Result<()> where Self: Sized

Sets the topic of a channel or requests the current one. If topic is an empty string, it won't be included in the message.

fn send_kill(&self, target: &str, message: &str) -> Result<()> where Self: Sized

Kills the target with the provided message.

fn send_kick(&self, chanlist: &str, nicklist: &str, message: &str) -> Result<()> where Self: Sized

Kicks the listed nicknames from the listed channels with a comment. If message is an empty string, it won't be included in the message.

fn send_mode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> where Self: Sized

Changes the mode of the target. If modeparmas is an empty string, it won't be included in the message.

fn send_samode(&self, target: &str, mode: &str, modeparams: &str) -> Result<()> where Self: Sized

Changes the mode of the target by force. If modeparams is an empty string, it won't be included in the message.

fn send_sanick(&self, old_nick: &str, new_nick: &str) -> Result<()> where Self: Sized

Forces a user to change from the old nickname to the new nickname.

fn send_invite(&self, nick: &str, chan: &str) -> Result<()> where Self: Sized

Invites a user to the specified channel.

fn send_quit(&self, msg: &str) -> Result<()> where Self: Sized

Quits the server entirely with a message. This defaults to Powered by Rust. if none is specified.

fn send_ctcp(&self, target: &str, msg: &str) -> Result<()> where Self: Sized

Sends a CTCP-escaped message to the specified target. This requires the CTCP feature to be enabled.

fn send_action(&self, target: &str, msg: &str) -> Result<()> where Self: Sized

Sends an action command to the specified target. This requires the CTCP feature to be enabled.

fn send_finger(&self, target: &str) -> Result<()> where Self: Sized

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_version(&self, target: &str) -> Result<()> where Self: Sized

Sends a version request to the specified target. This requires the CTCP feature to be enabled.

fn send_source(&self, target: &str) -> Result<()> where Self: Sized

Sends a source request to the specified target. This requires the CTCP feature to be enabled.

fn send_user_info(&self, target: &str) -> Result<()> where Self: Sized

Sends a user info request to the specified target. This requires the CTCP feature to be enabled.

fn send_ctcp_ping(&self, target: &str) -> Result<()> where Self: Sized

Sends a finger request to the specified target. This requires the CTCP feature to be enabled.

fn send_time(&self, target: &str) -> Result<()> where Self: Sized

Sends a time request to the specified target. This requires the CTCP feature to be enabled.